~ Commented some unused code # Fixed a crapton of warnings about type conversions # Fixed a crapton of warnings about signed/unsigned comparisons ~ Fixed some calls to setHWord and setWord had wrong order in parameters # Fixed some redefinition warnings on the system ~ Updated control scheme for game mode to a much usual one: D => square, S => cross, A => circle, W => triangle, arrow keys => pad

+++ include/crash/entry_cache.h
@@ -104,8 +104,8 @@ bool entryCache<T>::get(entry *key, T *&data)
   unsigned long EID        = key->EID;
   unsigned long cacheIndex = (EID >> 15) & 0xFF;
   
-  int newIndex = 0;
-  for (int lp = cacheIndex; lp < cacheIndex+0x100; lp++)
+  unsigned long newIndex = 0;
+  for (unsigned long lp = cacheIndex; lp < cacheIndex + 0x100; lp++)
   {
     newIndex = lp % 0x100;
